+++ packages/components-js/tests/e2e/specs/link.e2e.ts
@@ -1,10 +1,8 @@
 import {
   addEventListener,
-  expectedStyleOnFocus,
   getActiveElementId,
   getBrowser,
   getLifecycleStatus,
-  getOutlineStyle,
   initAddEventListener,
   selectNode,
   setContentWithDesignSystem,
@@ -192,50 +190,6 @@ describe('link', () => {
       linkElement.blur();
     });
     expect(await linkHasFocus()).toBe(false);
-  });
-
-  describe('focus state', () => {
-    it('should be shown by keyboard navigation only for shadowed <a>', async () => {
-      await initLink();
-
-      const link = await getLink();
-      const hidden = expectedStyleOnFocus({ color: 'transparent' });
-      const visible = expectedStyleOnFocus({ color: 'contrastHigh' });
-
-      expect(await getOutlineStyle(link)).toBe(hidden);
-
-      await link.click();
-
-      expect(await getOutlineStyle(link)).toBe(hidden);
-
-      await page.keyboard.down('ShiftLeft');
-      await page.keyboard.press('Tab');
-      await page.keyboard.up('ShiftLeft');
-      await page.keyboard.press('Tab');
-
-      expect(await getOutlineStyle(link)).toBe(visible);
-    });
-
-    it('should be shown by keyboard navigation only for slotted <a>', async () => {
-      await initLink({ useSlottedAnchor: true });
-
-      const link = await getSlottedLink();
-      const hidden = expectedStyleOnFocus({ color: 'transparent' });
-      const visible = expectedStyleOnFocus({ color: 'contrastHigh' });
-
-      expect(await getOutlineStyle(link, { pseudo: '::before' })).toBe(hidden);
-
-      await link.click();
-
-      expect(await getOutlineStyle(link, { pseudo: '::before' })).toBe(hidden);
-
-      await page.keyboard.down('ShiftLeft');
-      await page.keyboard.press('Tab');
-      await page.keyboard.up('ShiftLeft');
-      await page.keyboard.press('Tab');
-
-      expect(await getOutlineStyle(link, { pseudo: '::before' })).toBe(visible);
-    });
   });
 
   describe('lifecycle', () => {