+++ packages/components/src/components/content/text-list/text-list-item/text-list-item.tsx
@@ -1,6 +1,5 @@
 import { JSX, Component, Host, h, Element } from '@stencil/core';
-import { getAttribute, getTagName, insertSlottedStyles, throwIfParentIsNotOfKind } from '../../../../utils';
-import { P_ANIMATION_HOVER_DURATION } from '../../../../styles';
+import { getAttribute, throwIfParentIsNotOfKind } from '../../../../utils';
 import { addComponentCss } from './text-list-item-styles';
 
 @Component({
@@ -12,7 +11,6 @@ export class TextListItem {
 
   public connectedCallback(): void {
     throwIfParentIsNotOfKind(this.host, 'pTextList');
-    this.addSlottedStyles();
   }
 
   public componentWillRender(): void {
@@ -28,30 +26,5 @@ export class TextListItem {
         <slot />
       </Host>
     );
-  }
-
-  private addSlottedStyles(): void {
-    const tagName = getTagName(this.host);
-    const style = `${tagName} a {
-      color: inherit !important;
-      text-decoration: underline !important;
-      transition: color ${P_ANIMATION_HOVER_DURATION} ease !important;
-      outline: transparent solid 1px !important;
-      outline-offset: 1px !important;
-    }
-
-    ${tagName} a:hover {
-      color: #d5001c !important;
-    }
-
-    ${tagName} a:focus {
-      outline-color: currentColor !important;
-    }
-
-    ${tagName} a:focus:not(:focus-visible) {
-      outline-color: transparent !important;
-    }`;
-
-    insertSlottedStyles(this.host, style);
   }
 }