+++ packages/components/src/components/feedback/banner/banner-styles.ts
@@ -1,4 +1,13 @@
-import { addImportantToRule, attachCss, breakpoint, getCss, mediaQuery } from '../../../utils';
+import {
+  addImportantToRule,
+  attachCss,
+  breakpoint,
+  buildSlottedStyles,
+  getBaseSlottedStyles,
+  getCss,
+  insertSlottedStyles,
+  mediaQuery,
+} from '../../../utils';
 
 const easeOutQuad = 'cubic-bezier(0.5, 1, 0.89, 1)';
 
@@ -37,4 +46,12 @@ export const getComponentCss = (): string => {
 
 export const addComponentCss = (host: HTMLElement): void => {
   attachCss(host, getComponentCss());
+};
+
+export const getSlottedCss = (host: HTMLElement): string => {
+  return getCss(buildSlottedStyles(host, getBaseSlottedStyles()));
+};
+
+export const addSlottedCss = (host: HTMLElement): void => {
+  insertSlottedStyles(host, getSlottedCss(host));
 };