+++ packages/components/src/styles/internal.variables.ts
@@ -1,1 +1,0 @@
-export const P_ANIMATION_HOVER_DURATION = 'var(--p-animation-hover-duration, .24s)';