+++ packages/components/CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+#### Changed
+- Improved `:focus-visible` and `:hover:focus-visible` colors for `Link Social` and `Link`
+- Improved slotted `<a>` coloring in dark theme for `Link Social` and `Link`
+
+#### Fixed
+- Slotted `<a>` coloring in dark theme for `Text`, `Headline`, `Text List`, `Banner`, `Select Wrapper` and `Link Pure`
+
 ### [2.3.0] - 2021-07-28
 
 ### [2.3.0-beta.3] - 2021-07-28